Added symlinks restoration when disabling wine sandbox
Restored xdg keys order to match folder one

+++ lutris/util/wine/prefix.py
@@ -2,10 +2,11 @@
 import os
 from lutris.util.wine.registry import WineRegistry
 from lutris.util.log import logger
-from lutris.util import joypad, system
+from lutris.util import joypad, system, xdgshortcuts
 from lutris.util.display import DISPLAY_MANAGER
 
 DESKTOP_FOLDERS = ["Desktop", "My Documents", "My Music", "My Videos", "My Pictures"]
+DESKTOP_XDG = ["DESKTOP", "DOCUMENTS", "MUSIC", "VIDEOS", "PICTURES"]
 
 
 class WinePrefixManager:
@@ -72,7 +73,7 @@ class WinePrefixManager:
             return
         self.set_registry_key(key, dll, mode)
 
-    def desktop_integration(self, desktop_dir=None):
+    def desktop_integration(self, desktop_dir=None, restore=False):
         """Overwrite desktop integration"""
 
         user = os.getenv("USER")
@@ -84,19 +85,25 @@ class WinePrefixManager:
             desktop_dir = os.path.expanduser(desktop_dir)
 
         if system.path_exists(user_dir):
-            # Replace desktop integration symlinks
-            for item in DESKTOP_FOLDERS:
+            # Replace or restore desktop integration symlinks
+            for i,item in enumerate(DESKTOP_FOLDERS):
                 path = os.path.join(user_dir, item)
                 old_path = path + ".winecfg"
 
                 if os.path.islink(path):
-                    os.unlink(path)
+                    if not restore:
+                        os.unlink(path)
                 elif os.path.isdir(path):
                     try:
                         os.rmdir(path)
                     # We can't delete nonempty dir, so we rename as wine do.
                     except OSError:
                         os.rename(path, old_path)
+
+                if restore and not os.path.isdir(path):
+                    os.symlink(xdgshortcuts.get_xdg_entry(DESKTOP_XDG[i]),path)
+                    # We don't need all the others process of the loop
+                    continue
 
                 if desktop_dir != user_dir:
                     src_path = os.path.join(desktop_dir, item)
